chore(app): rename CopyPlaylistCommand to Command
